Separate button component & remove default text
