+++ global/supportedLanguages.json
@@ -10,7 +10,6 @@
         "displayName": "Display Name",
         "email": "Email",
         "findLures": "Find lures",
-        "addToMyLures":"Add to my lures",
         "addToMyLuresSucceed":"Lure has been added",
         "addToMyLuresDupe":"Lure already added",
         "addToMyLuresFail":"Something went wrong please try again",
@@ -155,7 +154,6 @@
         "displayName": "Nom d'affichage",
         "email": "Adress Courriel",
         "findLures": "Trouvez des leurres",
-        "addToMyLures":"Ajouter à mes leurres",
         "addToMyLuresSucceed":"Leurre ajouté",
         "addToMyLuresDupe":"Leurre déjà ajouté",
         "addToMyLuresFail":"Une erreur s'est produite. Veuillez réessayer",
